Flask/FlaskDB/requirements.txt
absl-py==1.3.0
alembic==1.9.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cvzone==1.5.6
cycler==0.11.0
dnspython==2.2.1
email-validator==1.3.0
Flask==2.2.2
Flask-Migrate==4.0.0
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
flatbuffers==23.1.4
fonttools==4.38.0
greenlet==2.0.1
idna==3.4
importlib-metadata==5.2.0
importlib-resources==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
Mako==1.2.4
MarkupSafe==2.1.1
matplotlib==3.5.3
mediapipe==0.9.0.1
mysql-connector-python==8.0.31
numpy==1.21.6
opencv-contrib-python==4.7.0.68
opencv-python==4.7.0.68
packaging==22.0
pandas==1.3.5
Pillow==9.3.0
protobuf==3.20.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.45
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
typing_extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.11.0
